audio_channel_out_mask_from_count: Return 5.1.4 position mask for 10 ch
When attempting to open a non-direct output stream with the 5.1.4
position channel mask, AudioFlinger fails and falls back to opening a
direct output stream instead. This is because
audio_channel_out_mask_from_count does not handle support for 10
channels. So, add the canonical position mask definition of 5.1.4 for 10
channels.

Test: m
Change-Id: I6437a9654d1b00be2feb828b0c81ec15a6d15986
Signed-off-by: Dmitrii <bankersenator@gmail.com>